// assets/battle_ts/model/XMatchData.ts
import { XCfgMapCfgItem, XCfgMapData } from "../xconfig/XCfgData";
import { XGameMode } from "../xconfig/XEnum";
import XPlayerModel from "./XPlayerModel";

export default class XMatchData {
    players: XPlayerModel[] = [];
    defenders: XPlayerModel[] = [];
    hunters: XPlayerModel[] = [];
    difficulty = 0;
    hideRoomIndex = 0;
    gameMode = XGameMode.E_Defense;

    mapCfg: XCfgMapCfgItem
    mapData: XCfgMapData
    mineUuid: string = ""
}